Write out the file input_data.csv so that the Cobalt prototype pipeline can read it.
The file will be written to out/input_data.csv. Here is a sample
of what the lines will look like:
9,EVELINA,Rancho Cucamonga,12,1
129,MARGARETT,Rochester,14,0
381,QUEENIE,San Bernardino,16,2
388,MARGARETT,Saginaw,12,7
583,MARGARETT,Rancho Cucamonga,11,10
516,KIMBERLI,Reno,8,5
853,MARGARETT,Roseville,10,6

The fields are:
user_id,module_name,city,hour_of_day,rating_out_of_10.

This does not include the search query yet. Usha is working on that.

Note that we don't bother having the straight counting pipeline read
this file. Instead the straight counting pipeline just uses the same
in-memory structure that is used to produce this file.

Change-Id: I9a97af41dc555e6070a284c8e12def40ffb58f8f