Use a power-law distribution for the cities and the module names.
We replace the current clipped normal distribution with a
power-law distribution. As suggested by mironov@ this gives us,
in his words, "a heavy head and a fat tail".

The new distribution will yield elements earlier in the static
list with greater frequency and so I moved the names of some
large cities to the front of the list so that the map looks
more real.

Change-Id: I25159db7dfe290b104e7a15d2fc0770729b06049