Adds the beginings of one of the randomizers.
We add the HelpQueryRandomizer that will extract the help search query
from each entry and then use Forculus threshold encryption to emit an
encrypted version of it. In this CL we are only adding the scaffolding--
we have not actually hooked up Forculus yet.

We have also refactored some of the existing code to make it more
reusable. We have created a utils directory with two utilities:
(i) file_util.py contains functions for reading and writing csv files
and also contains constants with the names of files and directories
that are shared between components.
(ii) data.py contains the Entry data type and functions for reading and
writing Entries. An Entry represents the synthetic input data.

We also updated the README.md with current instructions.

Change-Id: I6e1d4283c70fe5d5b3f586d5550411374007be20