Invoke forculus from the help_query_randomizer.
We replace the dummy implementation of the help_query_randomizer
with one that actually uses Forculus to write encrypted data.

We also include the code that should properly live in an anlayzer
that uses Forculus to do threshold decryption. This code will move
into an analyzer when more infrastructure is in place.

In the mean time it is possible to manually run this and see that
the output is correct. One can compare the contents of
help_query_analyzer_out.csv (the output of the Cobalt prototype
pipeline) with the contents of popular_help_queries.csv (the output
of the straight counting pipeline) to see that the Cobalt prototype
pipeline has correctly performed threshold encryption/decrption with
a threshold of 20.

I have written the exact counts of the helpstring into
help_query_analyzer_out.csv so that they can be compared to
those in popular_help_queries.csv. We may not want to write the exact
counts because it leaks some privacy perhaps.

I also slightly modified the public Forculus API.

Change-Id: I8aa27a909ff7e450b0f66d3b29a928e7f6dee37d