Improve the performance of Forculus encryption.
We bring the time to encrypt 10,000 entries down from over 10 seconds
to under 1 second.

Profiling indicated that the two most expensive operations were the
_ro_hmac function and the _unpack_bytes function. For _unpack_bytes
we provided a new faster implementation. For _ro_hmac we did the
following:

(i) We made a single global instance of the zeroed-out HMAC
object and whenever we need a new one we invoke copy() instead
of constructing a new one. Profiling indicated that the constructor
was expensive.

(ii) We allow the same instance of the HMAC object to be re-used
instead of creating a new one when doing so makes sense for
the computation. For example we use a single instance of HMAC
for each of the coefficients for a single polynomial rather than
using a new instance for each coefficient. We also use the same
instance to compute the initialization vector.

(iii) We cache the coefficient array, iv and ciphertext corresponding
to a plaintext so we never have to recompute them again for the same plaintext.

Change-Id: I88d55a72c7ab68561c4cad092f1e379053a441ca