Adds a help-query shuffler and analyzer.
We move the Forculus decryption invocation out of the help-query
randomizer and put it in the analyzer where it really belongs. We add the
necessary shuffler and analyzer infrastructure to make this work.

Note thate we intend to change the help-query example to a different
example that better shows off the usefulness of Forculus, but that will
happen in a later CL. For now we stick with the help-query nomenclature.

Change-Id: Ib143cf72cbde74ac721b4603358f028e1939782b