Keep the numerical fields of |Entry| typed as numbers upon reading.
Some of the fields of |Entry| are numbers at the time that an
|Entry| is created in generate_fake_data.py. Then we serialize
the list of Entries to a CSV file and the randomizers read them
again using data.py. At this point we should create an Entry with
the same numeric types for the fields. Previously we were reading
all the fields as strings.

Change-Id: I8457582262c70dc80ad007a406034eb6cd8169b8