Change _pack_into_bytes to big-endian.
This changes the implementation of _pack_into_bytes and one of the
side-effects of the change is to build a big-endian string instead
of a little-endian string. This does technically change the
implementation of Forculus but I believe in an unimportant way.

I want to make this change so that _pack_into_bytes is (more-or-less)
then inverse of _unpack_bytes. It was confusing to me that the
two functions used different endian-ness. There was no actual
bug though becuase nothing required that the functions were inverses.

Change-Id: I756147c7acc55ec79a225c9a5ecefaa98289121e