Starting to add some tests.
Now "cobalt.py test" will run an end-to-end test.
The end-to-end test first runs the pipeline the same
as "cobalt.py run" and then it invokes a script
to check the output. Currently the only check
being performed is to compare the output of the
Forculus pipeline to its counterpart in the
straight-counting pipeline.

Setting up this test actually already found a bug!
It turns out that in forculus.py in ComputeAndWriteResults()
we were ignoring the first line of the input file (I think because
in an earlier version the firs line was a header line).
I fixed this.

I also cleaned up a few miscellaneous things:
- I added a config file for Forculus where I put the threshold.
- I got rid of a couple of the printed messages that seemed to
clutter up the output more than add info. (I ended up removing
some print statements that Usha and Ananth added so if you
don't like the fact that I removed them and you want them
back just tell me and I'll put them back.)
- I switched the print statments about fastrand to log statements
so that if fastrand is installed you will not see the message
unless you pass the --verbose flag.

Change-Id: I523910c5171ca3db8ca06168fb88bcc7f98c7186