Removing quotes from city candidates.
The candidate and map file assumes quotes whereas the randomizer does not.
We choose not to have quotes for either; this CL fixes the candidates file.

If you have already tried to run RAPPOR on the existing map file
(city_map.csv) note that there might be a city_map.rda file in the same
directory generated by RAPPOR analysis. When dealing with large map files, the
R code chooses to "cache" the map file as a R data object. Delete that to see
the new candidate file and map file take effect.

Change-Id: Idc4f6221aff075fc8dbabaf7400e6709b5b9d403