Add url support for handling real world usecase for forculus.
This cl uses a mix of both low entropy urls such as the top 40 most
visited urls and some 5000 randomly generated high entropy urls that
are hard to crack such as Google doc urls. From the set of these high
entropy urls, 5 of them are marked as spammy, which appear 2% of the
time in the final outcome. These are generated with the following
randomness, which was purely through some experimentation:
   - 60% of the time a most-visited URL is used.
   - 38% of the time a uniformly random high-entropy URL is used.
   - 2% of the time one of the 5 spammy high-entropy URL is used.

This cl includes the following components: randomizer, shuffler and
analyzer for urls.

Also, refactored forculus randomizers and analyzers code.

Change-Id: Ie43d23a9c8d346f748ed4165ec77436ddf554dce