Adds a statistical dependence between city and rating.
Previously the distributions of city and rating were independent.
Now we have changed the distribution of rating to be a normal
distribution whose mean and variance depend on the city. The
more popular cities will have a higher mean and a smaller variance.

We played around with the parameters until the geo-chart looked
good. The goal was to have the biggest circles be green and the
smallest circles be red and the intermediate circles be yellow
and orange.

I also fixed the implementation of the skew parameter in
the function normalRandomInt(). That parameter had no
uses prior to this change.

I also moved a few cities that are real-life large cities
higher in our list of cities so that they would show up
on the map.

I also eliminated the cast to an integer for the average rating
in the straight-counting pipeline so that now a float is emitted.

Change-Id: Iad8de0fc7746df95d593363c39f5e3f0c9da42bd