Starting to add visualization for the Cobalt Prototype Pipeline.
We add a 95% confidence interval chart for the usage-by-module
pipeline.

Here is a screenshot of what the new chart looks like next
to the corresponding chart for the straight-counting pipeline.
https://screenshot.googleplex.com/RmJJ7WsUcMK.png

One thing you cannot see in the screenshot is the mouse-over text.
When you mouse over the colored regions in the left-side graph it
tells you the counts for each module in the straight-counting
pipeline. When you mouse over the blue dots in the chart on
the right-hand side it tells you the corresponding RAPPOR estimates
and then you can use the confidence intervals to see whether or not
the actual values fall within the confidence intervals.

We also organized visualization.html into a table where each
row has two columns: The left column is the straight-counting
pipeline and the right column is the corresponding
Cobalt prototype pipeline. Currently only the usage-my module
has the Cobalt Prototype Pipeline populated.

We also slightly modified the display of the top 50 help queries:
We changed the table to display all 50 queries instead of only
displaying 20 and requiring pagination.

Change-Id: If57a7409025e9e3cb3af3442d0d9b0662cd0f419