Includes the actual values in the confidence interval visualization.
See screenshot:
https://screenshot.googleplex.com/bNkh7enOGCi.png

We changes the look a little:
We use smaller blue circles for the estimate,
smaller red circles for the actual values and
grey line segments for the confidence intervals.

Notice that when the value does not appear at all
in straight-counting results the actual value is
rendered as zero.

Change-Id: Ia19425559e85013f4d477d5f278ce3d63382baf5