RAPPOR's decode.R: Simplify convoluted calculation.
In decode.R we replace a convoluted compution of the variance
of a binomial with an equivalent but more straight-forward
compution.

I have verified that my replacement is correct in three ways:

(1) The comment says that the variable |r| is supposed to be
the expectation of a reported 1. But that is just the sample
mean, i.e. v/N. So r should be equal to v/N.

(2) I added debugging statement to check if the new computation
returns the same values as the previous computation. The answer
is that the two computations do produce the same numbers except
in cases where the "clamp to [0, 1]" step of the previous
computation replaces a negative value with 0. But in this case
the old computation was encountering round-off error because
the clamping should not have been necessary. So the new
computation is superior.

(3) I verified algebraically that r = v/N. Here is the verification:

r = p_hats * p11 + (1 - p_hats) * p01

  = (v - p01 * N) * p11 / (N * p2) + p01 * (1 - (v - p01 * N) / (N * p2))

  = (v - p01 N )p11 + (p2N - v + p01*N)p01
   -----------------------------------
           p2 N

  = v(p11 - p01) + N * p01 * (p01 + p2 -p11)
   -----------------------------------
           p2 N

  = v (p2) +  N * p01 * (p2 - p2)
   -----------------------------------   becuase p2 = p11 - p01
           p2 N

  =  v/N

Change-Id: I84ecb83689433a29ca26efa409a666b4a61fd92f