Adds optional use of public key encrytpion.
We add the ability for the Randomizers to use the Analyzer's public key
to encrypt data so that the data is opaque to the shuffler. This is
enabled by default but can be disabled by passing --no-use-encryption
to cobalt.py run/test etc. The ability to disable the encryption
will be useful for debugging.

Currently this is only hooked up for Forculus. I will hook it up for
RAPPOR in a later CL.

Some notes
---------
I used a global variable _use_public_key_encryption in a couple
of files to record the state of the boolean. This is not an
ideal design but given our current architecture is was expedient.

I minimally changed the interface to Forculus. I simply added an
optional argument to the encode() and decode() which is a
function that can be used to encrypt just before writing and
decrypt just after reading.

I cleaned up one unrelated thing: we know invoke initializeFastRand()
just once instead of many times. I mostly did this because the function
prints a debug line and I didn't want to see it thousands of times.

Change-Id: Ic299f31a62153a855515be1f8a14ff7eb651c16f