Fixes the demo under --no-use-encryption
The demo was broken when --no-use-encryption was being passed.
This turned out to be because in analyzer.py we were not correctly
handling that case: we were telling an 'R' script to look for a file
called 'decrypted_shuffler_output_file.csv' but we were not producing
such a file. My fix was just to give the 'R' script a different file
path if we are not using encrytion.

In this CL I fix that problem but in the process of doing so I also
cleaned up some stuff that made it easier for me to understand what
was going on.

- In the function analyzeUsingRAPPOR() I replaced the first dictionary
  of parameters with real Python parameters. I left the second
  dictionary as a dictionary.

- I cleaned up the comments explaining what each parameter was for and
  what the requirements on it are.

- I removed a couple of constans from file_util.py because they were
  being used in only one file. If a file name is used in only one place
  we don't need to make it a global variable.

- I renamed some files so as to have consistent naming. This makes it
  easier to see related groups of files together:
   city_candidates.csv --> rappor_city_candidates.csv
   rappor_module_name_config.csv --> rappor_module_config.csv
   rappor_city_name_config.csv --> rappor_city_config.csv

Change-Id: I25c8043c8ee590dee4f1e07bd09fab477017cba5