Fixes help display for cobalt.py
We have a couple of global optional command-line
flags for cobalt.py
--verbose
--no_use_encryption

These flags are valid for all of the cobalt commands
such as cobalt.py run, cobalt.py test, cobalt.py analyze
etc.

I want to show the help string for --verbose and
--no_use_encryption both when the user types

python cobalt.py -h

and also when the user types

python cobalt.py run -h

In other words if the user types -h to get help I want
to show him the --verbose flag exists regardless of whether
or not he has also specified a sub-command.

This change does that. It is a tiny bit kludgy but
appears to be the only way to do it.

Change-Id: I926f95293a81d4ce92c194f960beb6e3f6c8aed5