Adds thin wrapper around OpenSSL/Boring SSL Crypto
We introduce a Cobalt-specific crypto library that is a thin wrapper
around OpenSSL/Boring. Its purposes are:

- To add a layer of indirection between Cobalt code and OpenSSL that
will give us flexibility for replacing the crypto layer if we need to.
- To add an easier to use interface with fewer option (because we have
chosen default options for Cobalt.)

We also add a GUnit test.

Because we have a real GUnit test now that links with Boring we no
longer have a need for our dummy utils and tests (they were there just
to test the build and testing system) so we delete them.

We are also including some unrelated general infrastructure in this CL:
- cpplint and a wrapper around it that allows us to invoke it using
"cobaltb.py lint"
- A template for producing a c++ copywrite header and header file guard
for the SublimeText editor.
These two pieces of infrastructure were used to help format this CL
and I would rather check them in now along with an example of their
use rather than create a separate CL where they are not being used.

Change-Id: I7cdedd2bedb1ee9f00af66f49e5fc8b03c4fa00c