Adds a thin wrapper around the OpenSSL/BoringSSL HMAC function.
We add mac.h and mac.cc. This again serves the purpose of adding
a layer of abstraction between Cobalt and the implementation.

It also exposes a slightly simpler interface because we have selected
HMAC-SHA256 and this has a known fixed tag length.

We add a new test in mac_test.cc and we add the new file
types.h and move the typedef for |byte| to it.

Change-Id: I80b9b5cdd936e1dca018a266a1400518b7b182f5