Initial skeleton code for Shuffler receiver service with an end to end local test added for grpc testing.
The proto added is an initial version and will get updated once the
proto syntax on encoder side is finalized.

On the build infrastructure for shuffler, grpc and protobuf3 compilers
are required along with their corresponding golang plugins. Grpc and
protobuf3 dependencies are already added through setup.sh script, and
the corresponding go plugins are checked in to the thirdparty/go folder
as part of this cl.

Also added the submodule initialization for protobuf GO and grpc-go.

Change-Id: If067cf0d909fdf6d7c0b9e966ec74ea94d35f4da