Create dockerfiles and add GCE commands to cobaltb
We create a base image for a cobalt system and an analyzer image that
uses it.  In future, a shuffler image will be create.

Currently the base image is rather "fat" (for simplicity) but it can
be optimized in future to include only what's needed.  E.g., it can be
based off busybox instead of ubuntu, etc.

The cobalt base image is an ubuntu system with the gRPC and protobuf3
libraries copied to it.

Change-Id: I93b285a72239c8c3e46eafcc3b1451fb06481860