Adds an interface for ForculusEncrypter and ForculusDecrypter.
The algorithms/encoder folder will hold interfaces to the algorithms
to be used in the Encoder.
forculus_encrypter.h contains the interface to Forculus.

The algorithms/analyzer folder will hold interfaces to the algorithms
to be used in the Analyzer.
forculus_decrypter.h contains the interface to Forculus.

The algorithms forculus folder contains the implementatio classes for
Forculus. So far it is just a dummy implementation. I will work on the
real implementation next.

In addition I made the following miscellaneous minor fixes as part of
this CL:
- I fixed the header guard in some .h files that I noticed were wrong.
- I added package names to the top-level proto files

Change-Id: I301fdd0c458da7bfe7bf2a5668430e9392515163