Adds an interface for RapporEncoder and BasicRapporEncoder.
The algorithms/encoder folder will hold interfaces to the algorithms
to be used in the Encoder.
rappor_encoder.h contains the interface to String and Basic RAPPOR.

The algorithms/rappor folder will contain the implementatio classes for
RAPPOR. So far it contains just a dummy implementation for the encoder
side.

I also renamed the fields of RapporConfig and BasicRapporConfig.
I replaced p,q,f etc with meaningful names and used "p", "q", "f" etc
in the comments.

Change-Id: I4fd27214a568cfc56abeadb4b3feb7f27852de89