Adds the function RandomBits().
This function takes a float p  in the range [0.0, 1.0] and returns
a byte in which each bit is set to 1 with probability p. This is
needed for RAPPOR encoding. See random.h, random.cc and random_test.cc.

We also fixed up some miscellaneous unrelated things we found
in other adjacent files.

Change-Id: I84a16d949cd659a21891a5f94dbf4ca6b5067998