Moves ObservationBatch from cobalt.proto to analyzer.proto.
If a message is only used as a parameter to a service then that message
should be defined in the same file as the service. ObservationBatch is
only used as a paramter to the Analyzer service.

The shuffler is a client of the Analyzer service and so in that capacity
it needs to import analyzer.proto. That is how the shuffler gets
ObservationBatch (so ObservationBatch should not be in any .proto files
in the shuffler.)

EncryptedMessage is different. It is not only part of ObservationBatch
which is sent from the Shuffler to the Analyzer, it is also part of
Envelope which is sent from the encoder to the Shuffler. For this reason
Encrypted Message neees to stay in cobalt.proto.

Change-Id: I3693bf8ef08bd68b760639536c055426e2634be8