Implements RAPPOR config validation in the Encoder.
When a RapporEncoder or a BasicRapporEncoder is constructed we parse and
validate the provided config and record whether or not it is valid.
If it is not valid then the call to Encode will return kInvalidConfig.

- We check that the IRR probablilites are between 0 and 1 and that the
  PRR probability has not been set or is set to 0.
- For Basic RAPPOR we check that there is at least one category and
  none of the categories are emtpy. We put the categories into a map.
- For string RAPPOR we check that num_bloom_bits, num_cohorts, and
  num_hashes are all positive and not absurdly large.

We also update the API to take a ClientSecret instead of a string client_id.

We add a unit test for all this.

Change-Id: Ic1de31fa1292ee65a4d659c27124f019d0c252a5