Adds ClientSecret.
ClientSecret is a move-only wrapper around an array of bytes containing
a 128-bit secret generated using our CSPRNG.

Clients should invoke GenerateNewSecret() once, and then permanently and
securely store the token returned from GetToken().

After that clients should reconstruct the ClientSecret using FromToken().

A ClientSecret is passed to the Encoder by the client and the Encoder
uses it whenever it needs a unique per-client secret.

This will replace the use of client_id currently in the ForculusEncoder
and RapporEncoder APIs.

Change-Id: Ic43fcbc45bde004175be8cb64c0be009c5c732ac