Introduce ObservationMetadata
We refactor Observation and Envelope and ObservationBatch.

In this new scheme
(i) The Process() method in the shuffler recieves an
EncryptedMessage
(ii) The decryption of the EncryptedMessage is an Envelope
(iii) The Envelope contains a Manifest and an EncryptedMessage
encrypted to the next recipient. In V 0.1 the next recipient will
always be the Analyzer and the EncryptedMessage contains an Observation
encrypted to the Analyzer.
(iv) The Manifest contains an ObservationMetadata. This allows the
Shuffler to do batch-and-forward based on metric_id
(v) ObservationBatch now has an ObservationMetadata.
(vi) When the shuffler sends an ObservationBatch to the Analyzer it
sends a homogeneous batch: all of the
encrypcted observations have the same metric_id and day_index.
(v) The Shuffler puts this ObservationMetadata into the ObservationBatch
along with the batch of EncryptedMessages.
(vi) The analyzer receiver now gets the meta-data with which to form the
BigTable row key from the ObservationMetadata in the ObservationBatch.
The Analyzer receiver does not need to decrypt the EncryptedMessages
before persisting them to BigTable.
Change-Id: Ia47c3dcb693c68a78d93d1fd3f80dab230ba3ada