Adds conversion functions between calendar date and day index.
We introduce the type CalendarDate and the functions
CalendarDateToDayIndex and DayIndexToCalendarDate.

The latter is implemented in a straightforward way.

The former is implemented twice:

In CalendarDateToDayIndexAltImpl() in datetime_util_test.cc we use a
straightforward implementation that may not be as portable as we
need it to be because it relies on the non-standard function timegm.

In CalendarDateToDayIndex() itself we use a harder-to-read
implementation that has the advantage of being portable.

There are extensive tests of both implementations and so it is possible
to switch to the alternate implementation if we want in the future.

Change-Id: I34b4fa898e72f7a702b538d026fa1c8edcce64f9