Introduces a registration system for EncodingConfigs.
- We add the file config/registered/registered_encodings.txt containing the
official set of registered EncodingConfigs.

- Recall that a decision was made that in V 0.1 we would not have an Admin Server backed by
a database but would instead use checked-in files as the official
source of truth for the registered config objects. This file is that
official source of truth for the EncodingConfigs.

- The other config objects will come in later CLs. The reason to treat
EncodingConfigs first and separately is that these config objects are
the only ones that are needed on both the Encoder and the Analyzer.

- The format of the file is the protocol buffer textual representation of a
|RegisteredEncodings| message which is defined in config/config.proto.

- We add the class EncodingRegistry in config.h. This is the class that
both an Encoder client and the Analzyer will use to discover the list
of registered EncodingConfigs. A client first invokes the function
FromFile() passing in the path to the config file. This yields an
instance of EncodingRegistry which may then be queried for the
registered encodings.

- We add some test config files in the folder config/test_files. These
are consumed by config_test.cc

- The utility config_gen.cc was used manually to help understand what
the text format to be used in the config files should look like.

Change-Id: Ie29a4a8bb30cd9e5faa946464e0165de94e1dea0