Implement BasicRapporEncoder::Encode()
- We implement the algorithm for encoding the IRR for Basic Rappor.
(We do not plan to support PRR in V 0.1 of Cobalt.)

- We add an extensive suite of tests for this including performing
a chi-squared test.

- We add the function SetRandomForTesting() to BasicRapporEncoder
so that we can use a deterministic RNG for tests.

- We factored the class DeterministicRandom out of
util/crypto_util/random_test and put it in a stand-alone
header file so we could use it in our new test.

- Rearranged CMake files to make it clear which libraries depend
on Boring SSL at compile time.

Change-Id: I075c4b5461dd2dad17bb93c6e4cf7c7ca09b877d