Adds MetricRegistry and ReportRegistry in addition to EncodingRegistry.
We refactored EncodingRegistry to be the template Registry<> in
config.h

This entailed eliminating config.cc and putting the templatized impl
into config.h

In encoding_config.h we redefine EncodingRegistry as a template
specialization of Registry. This will be used in both an Encoder client
and the Analyzer.

In metric_config.h we define MetricRegistry and ReportRegistry as
template specializations of Registry. These will be used only
in the Analyzer.

We eliminated the need for config.proto: We put RegisteredEncodings
into encodings.proto and RegisteredMetrics and RegisteredReports
into metrics.proto.

We extended config_test.cc to also test the two new Registry types.

Change-Id: I048cf6d85ec0116ea591f6a1395867b1f786bbc8