Introdudes ForculusAnalyzer.
A ForculusAnalyzer is used for the purpose of performing a single
Forculus analysis in the Cobalt analyzer.

It is a layer on top of the ForculusDecrypter. The ForculusDecrypter is
responsible for collecting together the observations corresponding to
a single ciphertext, whereas a ForculusAnalyzer contains many
ForculusDecrypters, one for each ciphertext it sees.

Some other miscellaneous stuff in this CL:

- ForculusEncrypter should have taken customer_id and project_id. We
rectify this oversight.

- use --logtostderr=1 when running gtests so that error log messages
can be seen

- I added a variant of Base64Encode() that takes a string as input. I needed
this to form error log messages.

Change-Id: I3ddaf803e9f005fed97f7f999ce33ff8ebc91ec8