Adds EncryptValue() to ForculusEncrypter
We add a layer above the the function ForculusEncrypter::Encrypt()
that accepts a |ValuePart|, serializes it to a string, and
then invokes Encrypt() on that string. In this way we have an
interface to ForculusEncrypter that knows how to encrypt blobs as
well as strings. The implementation of the Encoder API will invoke
this method.

Change-Id: I1219c8c36e6d92a8516381fafc7ce60e1854582b