Clean up build a bit
We do the following things:
(1) We fix a compiler warning in field_element_test.cc
(2) We get rid of a redundant find_package(Protobuf)
(3) We add QUIET to find_package(Protobuf)
(4) We add back in several other items to TO_SKIP_ON_PARTIAL_CLEAN
Without (4) a partial clean was not warking. After a partial
clean we would still do a full build with 597 steps including
building all of third_party (even though we were not deleting
the third_party directory under out.) In addition to taking a long
time this is especially anoying because building glog give clang
warnings. With (4) a partial clean works. After a partial clean
there are only 93 build steps and we do not rebuild third_party.

I would like to get rid of the clang warning messages for glog
but I cannot figure out how to do it.

Change-Id: If664ba5c9eb0841cd6daf7976ca804fb00f0ad05