Use day_index instead of observation_date for Forculus.
A CalendarDate is simply a user-friendly wrapper around a
day_index. Earlier I thought that an Encoder client would need
to deal directly with CalendarDates and that is why I had
the ForculusEncoder take a CalendarDate. But now it seems
there is no need for the Encoder client to deal with
CalendarDates because the Encoder will internally compute
the current day_index and so it works better to have the
ForculusEncode accept a day_index direclty.

Change-Id: I63e0c78fed5fdcde6f8b93e1873da53fe82a4d27