Encoder API implementation
Here is the latest version of the Encoder API, this time with a full
implementation and unit test.

The API has changed a little based on a fuller understanding of what is
required for the implementation and based on some comments from the last
round of API review.

As requested, a unit test is provided so you can see the API in action.

- The ProjectContext object encapsulates the customer_id and project_id
as well as the EncodingRegistry and the MetricRegistry for the
encoding configs and metrics in the project.

-The Encode*() methods no longer take an Observation* but instead
return a Result that contains a Status, an Observation, and
ObservationMetaData

A few other miscellaneous things in this CL
- I added some logging of errors to RapporEncoder
- I added some comments to client_secret.h

Change-Id: I9ce65b338eebe59d37ac7355bf80599f846827b5