Start computing the Forculus epoch index
This CL takes care of the following TODO:

// TUODO(rudominer) Compute epoch_index based on day_index.
unt32_t epoch_index = day_index;

That is, we were not really computing the epoch_index but rather we were
assuming that all epochs were DAY epochs and taking the day_index as the
epoch index. This TODO occurred in forculus_analyzer.cc and
forculus_encrypter.cc.

- We introduce the functions DayIndexToWeekIndex() and DayIndexToMonthIndex()
  in datetime_util.cc

- We add forculus_utils.h with the function EpochIndexFromDayIndex()

- We use the function EpochIndexFromDayIndex() in forculus_analyzer.cc and
  forculus encrypter.cc.

Change-Id: I5bf52579dba94cf7a1f2d2c69089b441e2cec549