Adding support for hybrid public-key encryption.
This CL adds a C++ interface for hybrid public-key encryption using
ECDH.

(The following documentation is also found in the class declaration.)
Public key = g^x in an elliptic curve group (NIST P256) represented in
X9.62 serialization with a compressed point repsentation.

Private key = x stored in bytes and interpreted as a big-endian number
(as in the interface of BN_bin2bn in boringssl)

Enc(public key, message):
   1. Samples a fresh EC keypair (g^y, y)
   2. Samples a salt
   3. Computes symmetric key = HKDF(g^xy, salt) with SHA256
   compression function
   4. Chooses a fresh nonce of size SymmetricCipher::NONCE_SIZE
   5. (Symmetric) encrypts message using SymmetricCipher::encrypt with
   key and nonce computed above into ciphertext
   6. Publishes (public_key_part, salt, nonce, ciphertext) as the
   hybrid ciphertext, where public_key_part is the X9.62 serialization
   of g^y

Dec(private key, hybrid ciphertext = (public_key_part, salt, nonce,
    ciphertext)):
   1. Computes symmetric key = HKDF(g^xy, salt) with SHA256
   compression function from private key (x) and public_key_part (g^y)
   2. (Symmetric) decrypts ciphertext using SymmetricCipher::decrypt
   with key and nonce

We add a PublicCipher class with a natural API. setPublicKey,
setPrivateKey, encrypt, and decrypt.

NOTE: For this CL we do not design or implement a schema to serialize
various components of the output of the hybrid encryption. As with
having a separate nonce and ciphertext component of the output of
symmetric encryption, we have four separate components (defined in the
encrypt() and decrypt() interfaces).

Change-Id: I601497acb2f095f896a887dd958ef14c7aa4ae95