This cl adds the dispatching logic for Shuffler i.e., When a new Encoder message arrives at the Shuffler, receiver stores it locally and a separate thread dispatcher wakes up every one hour or so (as configured in |frequencyInHours| param).
The dispatcher then either sends the stored messages if enough
volume is reached (as specified by |threshold| param), or filters
the messages based on |disposalAgeInDays| config param and saves them
back to teh store for the next dispatch attempt.

This cl also cleans up the memstore implementation and adds a new api
to get the count of observations for a given metadata key.

Adds unit tests for dispatcher/receiver and memstore along with fixing
the functional test {cobalt_test.go} for the new Shuffler config.

Change-Id: I8a45acdb57490365d52a1341a959f9ab72e781f1