Refactor: Extract some code from Rappor Encoder into a new file
This CL is in preparation for an upcoming CL in which
I introduce a Basic RAPPOR analyzer.

This is just a refactor
in which I extract some code from rappor_encoder and put
it into rappor_config_validator so that I can also use it in
basic_rappor_analyzer later.

Similarly I extract some code from
rappor_encoder_test and put it into rappor_test_utils so
that I can use it in basic_rappor_analyzer_test later.

Change-Id: I90a30d53b438b0fa9350b3f926a7660fb9e84d9a