Refactor Analyzer Reporter
I broke the files reporter.h and reporter.cc into two:

report_master and report_generator.

This separates two concerns and makes it easier to test. In particular
I want to be able to test report_generator without report_master.

More detail:

report_master is responsible for knowing the schedule for when reports
need to be generated and invoking report_generator each time a report
needs to be generated.

report_generator is only responsible for generating a single report at a time.

report_master may in the future become a stand-alone process. That is why
it is structured so that it has a funciton named report_master_main().
Currently this function is just invoked in a new thread.

report_generator on the other hand is just an object that one instantiates
and invokes GenerateReport().

Change-Id: Id507eda22e5303a1d820725d91ad583e04810946