Introducing Analyzer Store V2
Now that we have a better understanding of the requirements for the
Analyzer store we are introducing a new version of the interface.

We want to do this incrementally, from the bottom-up. So in this CL
we have not replaced the old Store interface but rather left the
old Store interface in place for now and introduced a new one. In this
CL we have only implemented the new interface with an in-memory store--
we have not yet written a BigTable-backed implementation for the new
interface.

The old interface is called cobalt::analyzer::Store and is defined
in store.h. The old in-memory implementation is called MemStore
and is defined in mem_store.h

The new interface is called cobalt::analyzer::store::DataStore and is
defined in data_store.h. The new in-memory implementation is called
MemoryStore and is defined in memory_store.h

The main differences are:
- The new interface has the notion of multiple columns. BigTable supports
multiple columns per row and we plan to make use of this in the Cobalt Analyzer.

- The new interface has the direct support for doing incremental reads of
a large range of rows where the client wants to consume the large range in
small chunks.

We also added a CMakeLists.txt file to the store directory rather than
building the store directory out of its parent directory.

Change-Id: I1b1e83bcc0c2f37ea074811a39a16620857d87d1