Introduces ObservationStore
ObservationStore is a layer on top of DataStore. To read and write
Observations the Analyzer will use the ObservationStore instead of
directly accessing the DataStore.

An ObservationStore is used for storing and retrieving Observations.
Observations are added to the store by the Analyzer Service when they
are received from the Shuffler. Observations are queried from the
store by ReportGenerator.

We also fix a small bug in memory_store.cc.

Change-Id: Ief6909125aeafa7c7a000fa37164520f40dc05f7