Make the parameter |max_rows| non-optional.
Previously the parameter |max_rows| was optional in the method
ReadRows(). The client could specify max_rows=0 and that was supposed
to mean that the client did not care to specify a maximum. But
this turns out to be not that useful and awkward to test because the
implementations of the interface all imposed some practical maximum.

The only reason for the parameter is it controls how big of a result
set should be allowed to accumulate in memory. So it makes more sense
to demand that the user specify something.

Change-Id: Iefddca4405bf2fd51af4ec0cde98f3fca7af6b9e