Adds BigtableStore
BigtableStore is our second implementation of the DataStore interface.
(The other one is MemoryStore.)

We also introduce several helper files for working with the Bigtable
Emulator. These are used only in a testing environment.

Because of the earlier work we did with templatized tests we don't need
new test code we only need two new instantiations of existing testing
templates. These may be found in bigtable_store_emulator_test.cc
and observation_store_emulator_test.cc.

We also fixed a bug in which we were erroneously using std:move() in
a loop.

Change-Id: Idfb6e2855f74e04d3bc7fc61e81da2efff7e3308