Some enhancements to the test harness.
We add two new features:

(1) We introduce a new category of tests called BT Emulator
tests. These are tests that expect that an instance of the
Bigtable Emulator is running on the localhost at the default
port. The test harness will start and stop the emulator.

(2) We introduce flags that allow one to run subsets of
the tests. For example one can run only the BT Emulator tests
or everything except the BT Emulator tests.

Change-Id: Ia20ca9549a7eb6e94edbee5d9c9d891657995b7f