Refactor some Cobalt Protos
Rather than the Encoder client sending the Shuffler an encrypted Envelope containing
the encryption of a *single* Observation at a time, now the Encoder client sends the
Shuffler an encrypted Envelope containing many Observations. The Observations are
organized into ObservationBatches (all the Observations in an ObservationBatch
share the same meta-data) and an Envelope now holds many ObservationBatches.

We also moved some of the protos into new files:

- We moved ObservationMetadata and ObservationBatch into observation.proto. They are
easier to find there.

- We also moved Envelope into observation.proto since it is easier to understand in
that context.

We also eliminated the unneeded ShufflerResponse message and replaced it with google.protobuf.Empty.

Change-Id: Ie83551ffef0914b0f75d6d2c11a3e5cd8fcde25d